客户列表新增黑名单功能

+++ linkwe-api/src/main/java/com/linkwechat/controller/WeMomentsTaskController.java
@@ -103,7 +103,7 @@ public class WeMomentsTaskController extends BaseController {
     @ApiOperation("预估客户数量")
     @PostMapping("/estimate/num")
     public AjaxResult getEstimateCustomerNum(@Validated @RequestBody WeMomentsTaskEstimateCustomerNumRequest request) {
-        Integer integer = weMomentsCustomerService.estimateCustomerNum(request);
+        long integer = weMomentsCustomerService.estimateCustomerNum(request);
         return AjaxResult.success(integer);
     }
