+++ linkwe-service/src/main/java/com/linkwechat/service/IWeMomentsCustomerService.java
@@ -27,7 +27,7 @@ public interface IWeMomentsCustomerService extends IService<WeMomentsCustomer> {
      * @author WangYX
      * @date 2023/06/07 17:01
      */
-    Integer estimateCustomerNum(WeMomentsTaskEstimateCustomerNumRequest request);
+    long estimateCustomerNum(WeMomentsTaskEstimateCustomerNumRequest request);
 
     /**
      * 预估客户
